Introduce Plugins Base classes
We have a lot of base plugin entities: Context, Scenario, SLA and etc.
Sometimes names of plugins for different types can be equal. For example,
we can have ceilometer client and ceilometer context. It is a not bad
idea to have simple same names for different entities, so lets start support
it.

Current list of plugin bases:
 - rally.task.context.Context
 - rally.task.scenario.Scenario
 - rally.task.types.ResourceType
 - rally.task.exporter.TaskExporter
 - rally.task.processing.charts.Chart
 - rally.task.runner.ScenarioRunner
 - rally.task.sla.SLA
 - rally.deployment.serverprovider.provider.ProviderFactory
 - rally.deployment.engine.Engine
 - rally.osclients.OSClient

Support of multiple plugins with the same name in one plugin base is
restricted.

Aslo, this patch :
 - removes configure decorator from several plugin base classes, since they
   can not be used outside and provide just an interface for further plugins.

Change-Id: I82852986504b55ad8a0399ffaadf25927caedf3e