Allow max_avg_sla per atomic actions
This adds a plugin to collect the avg time duration for each
atomic action. With this, SLAs setting time maximums can be
applied to individual atomic actions that are part of a run.

For example:

 NeutronNetworks.create_and_list_ports:
    -
      sla:
        max_avg_duration_per_atomic:
          neutron.list_ports: 5.5

This would put a 5.5 second limit on the average amount of
time spent listing ports. During the create_and_list_ports
test run.

This allows projects to put more fine-grained restrictions
on different types of operations (e.g. GET vs POST).

Change-Id: I752a53ae3b2f47a05057ad8d4688f3f74d75ecf0