Validate scenario arguments
We have a lot of validators for scenarios, but we do not check input
arguments. There are two cases:
1) user forget specify some argument;
2) user specify redundant argument

In both cases, we will get "TypeError" for each iteration.

This validator is not implemented like others, since it doesn't need
constructed clients and it looks like more syntax check.

Also, this patch fixes one of our samples:
    samples/tasks/scenarios/dummy/dummy-random-action.[json/yaml]

Change-Id: Id3aad571dfc93f8074c724595440979cfd435e2c