disable urllib3 warnings only if the library provide them
SNIMissingWarning was introduced in urllib3 in 1.13.

When using rally with an older version of urllib3 (f.e. as used in
liberty, urllib3>=1.8.3), the SNIMissingWarning may be not present, in
which case the code will crash with AttributeError.

So first check whether the warning is known to the urllib3 library being
used with rally before trying to disable it.

Change-Id: Idcdc7b43a032212064c4ed2e1a69b67656bf801c
Closes-Bug: #1573650