[Verify] Add tempest configs and move to new section
Moved the old image and role options to the [tempest] section and added a
few flavor related config options. With these new options it is possible
to configure the RAM size of the two flavors used when running tempest
and the RAM size of the orchestration instance type flavor.

These were previously hardcoded to 64M, 128M and 64M respectively which
may cause the test image to fail under certain conditions, due to lack of
memory (e.g. AArch64 VMs seem to fall into this category).

After applying this fix, the three ram sizes and all other tempest related
options can be configured in the [tempest] section of /etc/rally/rally.conf.

Closes-Bug: #1580695

Change-Id: I33b8682310a42e66b408802a0d955fc886f6c8ed
Signed-off-by: Ciprian Barbu <ciprian.barbu@gmail.com>