[Verify] Fixing sahara plugin tempest tests
If you take a look at [1], you will see that 2 sahara plugin tempest tests
fail. It is because of not correctly configured the "plugins" option in the
[data_processing-feature-enabled] group. We have a hard-coded list of plugins
there. It looks like the best solution, that will be suitable for all envs,
is not to configure this option at all and rely on the default values.

[1] http://logs.openstack.org/32/310232/13/check/gate-rally-dsvm-verify-full-create-resources/a1a3b9d/rally-verify/9_verify_results.html.gz

Change-Id: I682e1a9fa2703a1d58ecf92b5834820095d01d65