Add default role name to rally config
If the OpenStack env does not have a 'member' role, the
rally task will get a WARNING: Unable to set member role
to created user. And then task will stop due to the 401
Unauthorized error.

This patch adds default role to rally config. So for each,
test, rally will create the test user with the set
keystone_default_role.

Closes-Bug: #1595081

Change-Id: Ic97ce50f40d3a3e7f9e8fc6ef142c5465ab41a51