Fix output of small floats in results
For now we have output of small floats in scientific notation:
3.79085540771e-05
Fix that to 0.000037. Also floats in fixed place will output as:
  0.0000001 -> 0.0
  0.000000 -> 0.0
  37 -> 37.0
  1.0000001 -> 1.0
  1.0000011 -> 1.000001
  1.0000019 -> 1.000002

This can be made in different ways like:
  "%f" % float(decimal.Decimal(num).normalize())
or in more weird way but I choose that one (see in the patch) like
the most flat, simple and explicit.

Change-Id: I4b9cdbb01f3f8e2eeaedc9d33e1203247ea79f89