[existing users] Restore original quotas
Behaviors of quotas cleanup should be different in case of existing users and
new ones(created by users context).

In case of existing users, we should not delete quotas, since we should not
change user pre-defined data.
In case of new users, we can't do the same as with existing users. Restore
default quotas means that services databases will continue storing quotas even
after tenant removement. So we need to remove quotas in this case.

Closes-Bug: #1595578

Change-Id: I6d42942f7fcdc67f221dc737495663eec065c8e4