[Verify] Updating Tempest config file
Some tests (for boto, horizon, etc.) were removed from Tempest
and now there is no need to keep the corresponding  options in
Tempest config file.

Some options in Tempest were moved from one section to another and we
should to do the corresponding changes in Rally to be up to date with
the latest Tempest version.

Change-Id: Ied4ff74c6c8edc7fe30c0e891c668a5b9d390fb4