Port of Auth and Desicnate to new style
Since current scenario implementation transforms method
to class at runtime, it is reasonable to have class-based
scenario implementation which will be much simpler to use
behind the scenes.

This class should be based on Scenario and do not break
compatibility. The only required method is run() which
is actually a body of scenario.

Change-Id: I2fb168c9762311975f1a9fad1412061944ee9a5d