[Reports] In charts, fix setting red color for non-error metric
There is a bug in charts StackedArea and Lines, which appears
in case if there are a >= 9 metrics - in this case metric #9
gets red color, which is also set for failed_duration.
As result, tehere are 2 metrics with same color on chart.

This fixes the bug by introducing standard way of color
generation for all charts, which skips failure-color (red)
when generating next color.

Change-Id: Ideaed933f3fdb5b3465124fc5b8ae730119cb08c