Moving of Heat scenarios into separate classes
Since current scenario implementation transforms method to class at runtime,
it is reasonable to have class-based scenario implementation which will be
much simpler to use behind the scenes.

This class should be based on Scenario and do not break compatibility.
The only required method is run() which is actually a body of scenario.

This patch contains changes for HEAT scenarios and tests

Change-Id: I92a1b63c40a67abbe05a78ba35ca53dc73952e75