Moving of Neutrone scenarios into separate classes
Implements blueprint class-based-scenarios

Since current scenario implementation transforms method to class
at runtime, it is reasonable to have class-based scenario implementation
which will be much simpler to use behind the scenes.

This class should be based on Scenario and do not break compatibility.
The only required method is run() which is actually a body of scenario.

This patch contains changes in following modules
(both scenarios and tests):
Neutron

Change-Id: Ib7284d475e2e2f792f0dc39a176249b0e6ca2560