Port all Sahara scenarios to new style
Implements blueprint class-based-scenarios

Since current scenario implementation transforms method to class at runtime,
it is reasonable to have class-based scenario implementation which will be
much simpler to use behind the scenes.

This class should be based on Scenario and do not break compatibility. The
only required method is run() which is actually a body of scenario.

This patch contains changes in following components
(both scenarios and tests:
   Sahara

Change-Id: I7b1437e52efebcdacce29792d5ec51fbd496e03a