[Verify] Fixing issue with downloading images for the tests
Scenario tests in Tempest require an image file and Rally always
downloads the image file by the link specified in CONF.tempest.img_url
regardless of CONF.tempest.img_name_regex is set or not.

If CONF.tempest.img_name_regex is set, we should find an image matching
to the regex in Glance and download it for the tests.
If CONF.tempest.img_name_regex is not set (or we didn't find the image
matching to CONF.tempest.img_name_regex), we should download the image
by the link specified in CONF.tempest.img_url.

Also, this patch changes Tempest plugin URL in the rally_verify.py file from

  https://github.com/MBonell/hello-world-tempest-plugin

to

  https://git.openstack.org/openstack/ceilometer

to resolve jenkins gate issues with rally verify jobs.

Tempest plugin is inside the ceilometer project [1].

[1] https://github.com/openstack/ceilometer/tree/master/ceilometer/tests/tempest

Change-Id: I3210e04bb53e77acab3a6bec172bb8d035bb8af2